Update README.md
The website is a travel blog where I record all my trips and the destinations I have been to in each country. The website provides a summary about the countries and the cities that were visited supported with images and pictures that were taken during that trip. The website also provides a historical summary of the main tourism destination to not only let the user enjoy the trip and the pictures but also gain some information about the country and the city that were visited.